+++ app/scripts/services/dimManifestService.factory.js
@@ -4,9 +4,9 @@
   angular.module('dimApp')
     .factory('dimManifestService', ManifestService);
 
-  ManifestService.$inject = ['$q', 'dimBungieService', '$http', 'toaster', 'dimSettingsService', '$translate'];
+  ManifestService.$inject = ['$q', 'dimBungieService', '$http', 'toaster', 'dimSettingsService', '$translate', '$rootScope'];
 
-  function ManifestService($q, dimBungieService, $http, toaster, dimSettingsService, $translate) {
+  function ManifestService($q, dimBungieService, $http, toaster, dimSettingsService, $translate, $rootScope) {
     // Testing flags
     const alwaysLoadRemote = false;
 
@@ -144,6 +144,7 @@
             })
             .catch((e) => console.log('Error saving manifest file', e));
 
+          $rootScope.$broadcast('dim-new-manifest');
           var typedArray = new Uint8Array(arraybuffer);
           return typedArray;
         });