+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * We swapped in "special ammo synth" and "primary ammo synth" instead of "motes of light" and "strange coins" for the farming mode quick gather buttons. They seemed more useful in the heat of battle.
 * When dequipping an item, we try harder to find a good item to equip in its place. We also prefer replacing exotics with other exotics, and correctly handle The Life Exotic perk.
 * Lots of new translations and localized strings.
+* Vendors update when you reach a new level in their associated faction, or when you change faction alignment.
 
 # 3.14.1